requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pytest==6.2.5
requests==2.26.0
python-dateutil==2.8.2
pydantic==1.10.2